Distructured container components for all pages
