+++ src/pages/Experience.js
@@ -1,15 +1,36 @@
 import React from "react";
-import { ExperienceContainer, Container } from "../components/Container";
+import { Container } from "../components/Container";
 import ExperienceList from "../components/Experience/ExperienceList";
 import { TextBig, TextLarge } from "../components/Text";
+import styled from "styled-components";
+import { breakpoints } from "./../helpers/Media";
+
+export const ExperienceContainer = styled(Container)`
+  @media screen and (max-width: ${breakpoints.screen4k}) {
+    padding: 100px 300px 50px 300px;
+    grid-template-rows: 50px 1fr 300px;
+    grid-template-columns: 1fr;
+  }
+  @media screen and (max-width: ${breakpoints.laptopMScreen}) {
+    padding: 100px 200px 50px 200px;
+    grid-template-rows: 50px 200px 1fr 200px;
+  }
+  @media screen and (max-width: ${breakpoints.laptopScreen}) {
+    padding: 100px 150px 50px 150px;
+    grid-template-rows: 50px 1fr 500px;
+    grid-template-columns: 1fr;
+  }
+  @media screen and (max-width: ${breakpoints.tabletScreen}) {
+    padding: 100px 100px 50px 100px;
+  }
+  @media screen and (max-width: ${breakpoints.mobileLScreen}) {
+    padding: 100px 30px 50px 30px;
+  }
+`;
 
 export default function Experience() {
   return (
-    <ExperienceContainer
-      gap="50px"
-      id="experience"
-      padding="100px 0px 100px 0px"
-    >
+    <ExperienceContainer gap="40px" id="experience">
       <Container place="center start">
         <TextBig>Experience</TextBig>
       </Container>
