+++ src/pages/Footer.js
@@ -13,8 +13,13 @@ const PaddingFooter = styled(Container)`
   @media screen and (max-width: ${breakpoints.screen4k}) {
     grid-template-columns: repeat(2, 1fr);
     grid-templete-rows: 1fr;
+    padding: 100px 300px 50px 300px;
+  }
+  @media screen and (max-width: ${breakpoints.laptopMScreen}) {
+    padding: 100px 200px 50px 200px;
   }
   @media screen and (max-width: ${breakpoints.laptopScreen}) {
+    padding: 100px 150px 50px 150px;
     grid-template-rows: repeat(2, 1fr);
     grid-template-columns: 1fr;
     place-items: center;
@@ -22,15 +27,17 @@ const PaddingFooter = styled(Container)`
   }
   @media screen and (max-width: ${breakpoints.tabletScreen}) {
     place-items: center;
+    padding: 100px 100px 50px 100px;
   }
   @media screen and (max-width: ${breakpoints.mobileLScreen}) {
     place-items: center;
+    padding: 100px 30px 50px 30px;
   }
 `;
 
 export default function Footer() {
   return (
-    <PaddingFooter padding="50px 0px 10px 0px">
+    <PaddingFooter>
       <Container place="center start">
         <TextRegular textAlign="center">
           Petros Chantzopoulos - April 2021 - v3.0.0
