+++ src/pages/Skills.js
@@ -1,12 +1,35 @@
 import React from "react";
-import { Container, SkillsContainer } from "./../components/Container";
+import { Container } from "./../components/Container";
 import SkillsTopicList from "../components/Skills/SkillsTopicList";
 import SkillsStack from "../components/Skills/SkillsStack";
 import { TextLarge, TextBig } from "../components/Text";
+import styled from "styled-components";
+import { breakpoints } from "./../helpers/Media";
+
+export const SkillsContainer = styled(Container)`
+  @media screen and (max-width: ${breakpoints.screen4k}) {
+    padding: 100px 300px 50px 300px;
+    grid-template-rows: 50px 150px 1fr 200px;
+    grid-template-columns: 1fr;
+  }
+  @media screen and (max-width: ${breakpoints.laptopMScreen}) {
+    padding: 100px 200px 50px 200px;
+    grid-template-rows: 50px 200px 1fr 200px;
+  }
+  @media screen and (max-width: ${breakpoints.laptopScreen}) {
+    padding: 100px 150px 50px 150px;
+  }
+  @media screen and (max-width: ${breakpoints.tabletScreen}) {
+    padding: 100px 100px 50px 100px;
+  }
+  @media screen and (max-width: ${breakpoints.mobileLScreen}) {
+    padding: 100px 30px 50px 30px;
+  }
+`;
 
 export default function Skills() {
   return (
-    <SkillsContainer gap="40px" id="skills" padding="100px 0px 100px 0px">
+    <SkillsContainer gap="40px" id="skills">
       <Container place="center start">
         <TextBig textAlign="left">Skills</TextBig>
       </Container>
